Adjust spec files for augmented graph generation
When syncing back up with the CSA tooling, I noticed that the CSA parser
does not allow assignment graph files with non-contiguous node numbers.
This doesn't appear to be part of the DIMACS graph spec, but it is also
non-trivial to modify the original CSA code to accept this.

It should be straightforward to modify the augmented graph generation
code to output contiguous node numbers.

These spec changes update our augmented graph fixtures to use contiguous
node numbers.

Downstream, we will probably need to supply the number of source nodes in
the original (non-augmented) graph when we convert the augmented flow to
a non-augmented flow later.